Add support for `cpputestTestAdapter.preLaunchTask` setting     (#61)
* run(): RunAllTests() if the only test is named 'error'

This can happen if a previous TestLoadFinishedEvent contained an
errorMessage instead of a suite.

No code takes advantage of this yet. It is being added in preparation
for preLaunchTask support, which will emit such an event if the task
fails.

* Add support for `cpputestTestAdapter.preLaunchTask` setting

The configured task will be run at the beginning of load(), run(), and
debug(). This can be used to rebuild the test executable.

Addresses issue #42

* Bail out of run() or debug() if preLaunchTask fails